+++ lib/src/model/document_result.dart
@@ -5,8 +5,9 @@ class DocumentResult {
   DocumentSideResult? back;
   String? typeSelected;
   String? countrySelected;
+  String? nfcMediaId;
 
-  DocumentResult({this.front, this.back, this.typeSelected, this.countrySelected});
+  DocumentResult({this.front, this.back, this.typeSelected, this.countrySelected, this.nfcMediaId});
 
   @override
   bool operator ==(Object other) {
@@ -15,11 +16,13 @@ class DocumentResult {
     if (other.back != back) return false;
     if (other.typeSelected != typeSelected) return false;
     if (other.countrySelected != countrySelected) return false;
+    if (other.nfcMediaId != nfcMediaId) return false;
     return true;
   }
 
   @override
-  int get hashCode => front.hashCode + back.hashCode + typeSelected.hashCode + countrySelected.hashCode;
+  int get hashCode =>
+      front.hashCode + back.hashCode + typeSelected.hashCode + countrySelected.hashCode + nfcMediaId.hashCode;
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = <String, dynamic>{};
@@ -31,6 +34,7 @@ class DocumentResult {
     }
     data['typeSelected'] = typeSelected;
     data['countrySelected'] = countrySelected;
+    data['nfcMediaId'] = nfcMediaId;
     return data;
   }
 }