+++ lib/src/model/face_capture_type.dart
@@ -1,1 +1,1 @@
-enum FaceCaptureType { photo, video }
+enum FaceCaptureType { photo, video, motion }