+++ lib/src/model/media_callback.dart
@@ -1,0 +1,5 @@
+import 'package:onfido_sdk/src/model/media_result.dart';
+
+abstract class OnfidoMediaCallback {
+  void onMediaCaptured({required OnfidoMediaResult result});
+}