+++ lib/src/serializer/serializer.dart
@@ -1,3 +1,4 @@
 export 'package:onfido_sdk/src/serializer/start_options_serializer.dart';
 export 'package:onfido_sdk/src/serializer/start_studio_serializer.dart';
 export 'package:onfido_sdk/src/serializer/onfido_result_serializer.dart';
+export 'package:onfido_sdk/src/serializer/onfido_media_result_serializer.dart';